Add .editorconfig file
.editorconfig files are widely supported and offer basic configuration
 options for the project.

This sets the indention style, line-ending, character set,  tells the
editor to make sure there's a newline at the end of the file and to
strip trailing whitespace.

For directories within the coreboot directory that would prefer a
different setting, additional  .editorconfig files can be placed in
those directories to override any of these settings.

See the EditorConfig website for more information.
https://web.archive.org/web/https://editorconfig.org

Change-Id: Iecf1c5450edb0db533569189aa45233b91997870
Signed-off-by: Martin Roth <martin@coreboot.org>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/35185
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-by: Patrick Georgi <pgeorgi@google.com>